Point2D is double. Added android logic to handle response from face recognition.

+++ Bert/common/src/main/kotlin/chuckcoughlin/bert/common/model/Point2D.kt
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024. Charles Coughlin. All Rights Reserved.
+ * Copyright 2024-2025. Charles Coughlin. All Rights Reserved.
  * MIT License.
  */
 package chuckcoughlin.bert.common.model
@@ -7,7 +7,7 @@ package chuckcoughlin.bert.common.model
 /**
  * Use our own class of Point to make it easier to serialize and deserialize.
  */
-data class Point2D (val x1:Float, val x2:Float) {
-    val x: Float = x1
-    val y: Float = x2
+data class Point2D (val x1:Double, val x2:Double) {
+    val x: Double = x1
+    val y: Double = x2
 }
